+++ activestorage/test/test_helper.rb
@@ -130,6 +130,7 @@ class User < ActiveRecord::Base
   validates :name, presence: true
 
   has_one_attached :avatar
+  has_one_attached :avatar_with_prefix, prefix: -> (record, _attachment) { record.name }
   has_one_attached :cover_photo, dependent: false, service: :local
   has_one_attached :avatar_with_variants do |attachable|
     attachable.variant :thumb, resize_to_limit: [100, 100]
@@ -147,6 +148,7 @@ class User < ActiveRecord::Base
   has_one_attached :name_pronunciation_audio
 
   has_many_attached :highlights
+  has_many_attached :highlights_with_prefix, prefix: -> (record, _attachment) { record.name }
   has_many_attached :vlogs, dependent: false, service: :local
   has_many_attached :highlights_with_variants do |attachable|
     attachable.variant :thumb, resize_to_limit: [100, 100]